# apps/backend/user/user-app-api/contracts/openapi/user-app-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/agritech/catalog": {
      "get": {
        "operationId": "ProductController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductListDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Bad Request"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [400],
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [401],
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Not Found"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [404],
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Internal Server Error"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [500],
                      "example": 500
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "nrb.sid": []
          },
          {
            "__Host-nrb.sid": []
          }
        ],
        "tags": ["agritech-catalog"]
      }
    },
    "/agritech/catalog/{id}": {
      "get": {
        "operationId": "ProductController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductViewDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Bad Request"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [400],
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [401],
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Not Found"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [404],
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Internal Server Error"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [500],
                      "example": 500
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "nrb.sid": []
          },
          {
            "__Host-nrb.sid": []
          }
        ],
        "tags": ["agritech-catalog"]
      }
    }
  },
  "info": {
    "title": "user-app-api",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nrb.sid",
        "description": "Development/default HTTP session cookie. SESSION_COOKIE_NAME may override the runtime name."
      },
      "__Host-nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-nrb.sid",
        "description": "Production default HTTPS session cookie. SESSION_COOKIE_NAME may override the runtime name."
      }
    },
    "schemas": {
...
      "ProductViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "nameRu": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": ["fertilizer", "seed", "pesticide", "equipment", "irrigation", "other"]
          },
          "description": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          },
          "priceUzs": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "number"
          },
          "region": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "out_of_stock"]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "description",
          "supplierName",
          "priceUzs",
          "unit",
          "stockQuantity",
          "region",
          "status",
          "images",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductViewDto"
            }
          }
        },
        "required": ["items"]
      }
    }
  }
}